chore(wallet-config): FIle reorganisation (extract types)
